backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
gunicorn
